+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Gracefully stopping the RabbitMQ communication background thread no longer
   prints an alarming error message.
 
+- Dead URLs are now presented in the error messages.
+
 ## Version 3.20.1, 22nd June 2023
 
 "Artificial CEO"